Return an error when the RBAC manager is started with the --manage flag
The --manage flag lets you change the RBAC management policy from Basic
to All. All was the default until v1.13, where we changed it to Basic.

The All policy starts an additional controller that manages namespace
scoped RBAC roles and bindings. I suspect no-one uses this policy. No
one seemed to notice when we switched the default policy to Basic.

Just in case, my plan to remove it is:

* Hide the --manage flag to make it less discoverable
* Make --manage return an error linking to https://github.com/crossplane/crossplane/issues/5227
* Add a --deprecated-manage flag as a 'break glass' option

The idea is to break anyone who cares about configuring the management
policy enough to notice the flag is going away, but not so much they
can't easily recover (for now) by using --deprecated-manage instead. My
hope is that if anyone actually uses the All policy they'll see the
error and comment on the issue.

Signed-off-by: Nic Cope <nicc@rk0n.org>